+++ web-ui/src/components/event-components/shared/ProgressButton.tsx
@@ -1,4 +1,4 @@
-import { Box, Fab } from '@mui/material';
+import { Fab } from '@mui/material';
 import ArrowBackIosNewRoundedIcon from '@mui/icons-material/ArrowBackIosNewRounded';
 import PlayArrowRounded from '@mui/icons-material/PlayArrowRounded';
 import SkipNextRounded from '@mui/icons-material/SkipNextRounded';