+++ web-ui/src/components/replay-components/ReplaySprintRaceFrame.tsx
@@ -1,6 +1,6 @@
 import { useMemo, useState } from 'react';
 import { useMutation } from '@tanstack/react-query';
-import { EventRecordsApi, EventDtoOutput } from '../../../generated';
+import { EventRecordsApi, EventDto } from '../../../generated';
 import { SprintEventRecordDtoOutput as EventRecordDto } from '../../../generated/models/SprintEventRecordDtoOutput';
 import { getRaceDurationBySize } from '../event-components/event-utils';
 import defaultConfig from '../../default-config';
@@ -8,7 +8,7 @@ import { SprintRaceUI } from '../event-components/sprint-race/SprintRaceUI';
 import { useReplayTickDelay } from '../../hooks/useReplayTickDelay';
 
 interface ReplaySprintRaceFrameProps {
-  event: EventDtoOutput;
+  event: EventDto;
   tick: number;
   maxTick: number;
 }